+++ src/skat/util.c
@@ -21,10 +21,10 @@ open_random_fd(void) {
   }
 }
 
-int
-util_rand_int(const int min, const int max) {
-  int random;
-  read(random_fd, &random, sizeof(int));
+size_t
+util_rand_int(const size_t min, const size_t max) {
+  size_t random;
+  read(random_fd, &random, sizeof(size_t));
 
   return (random % (max - min)) + min;
 }
